headless openMole
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@9144 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.headless/src/msi/gama/headless/common/ISimulator.java
@@ -20,5 +20,5 @@ public interface ISimulator {
     public DataType getVariableTypeWithName(java.lang.String name) ;
     public boolean containVariableWithName(java.lang.String name) ;
     public void load(String var, int exp, String expName);
-    public void setSeed(double seed); 
+    public void setSeed(long seed); 
 }
